Create LICENSE.md
This is an adopted license from another website, I went over a condensed list and it would allow Commercial use Distribution, Modification, Patent use, and Private use. The conditions to use our code would require the users to: Disclose source, License and copyright notice, and  State the changes they made. Also users who use the network would be allowed a copy of the source code and it would require them to release their modifications under the same license as the original.